Delay start to allow wifi to connect
Attempting to access the network before wifi has associated to an
AP and acquired an IP address will crash the wifi driver and require
a reboot of the system to recover.

Connman 1.31 introduced 'connmand-wait-online' and is probably the
right thing to use.  systemd-networkd-wait-online also exists and
might be an option for those not using connman, but are using
systemd.